Update bot service to improve error handling and logging, and enhance user data management
- Added `stats_data.json` to `.gitignore` to prevent tracking of runtime data.
- Refactored `handle_admin_reply` in `notification_service.py` to remove debug prints and improve exception handling.
- Cleaned up print statements in various bot service files to streamline logging and reduce clutter.
- Introduced a new method in `user_service.py` for case-insensitive username lookup.
- Updated schemas for service tickets and users to include optional `object_id` for better data management.
- Enhanced pagination and search functionality in user and service ticket APIs for improved user experience.

+++ infrastructure/services/web_service/src/api/schemas/rental_spaces.py
@@ -8,7 +8,7 @@ class SpaceResponse(BaseModel):
     id: Optional[int] = None
     object_id: int
     floor: str
-    size: int
+    size: float
     description: Optional[str] = None
     photos: List[str] = []
     status: str = "FREE"
@@ -22,7 +22,7 @@ class CreateSpaceRequest(BaseModel):
 
     object_id: int
     floor: str
-    size: int
+    size: float
     description: Optional[str] = None
     photos: List[str] = []
     status: str = "FREE"
@@ -34,7 +34,7 @@ class UpdateSpaceBody(BaseModel):
 
     object_id: Optional[int] = None
     floor: Optional[str] = None
-    size: Optional[int] = None
+    size: Optional[float] = None
     description: Optional[str] = None
     photos: Optional[List[str]] = None
     status: Optional[str] = None
